+++ pages/admin/permit-holders.tsx
@@ -45,7 +45,7 @@ import { formatDate } from '@lib/utils/format'; // Date formatter util
 import SetPermitHolderToInactiveModal from '@components/admin/permit-holders/modals/SetPermitHolderToInactiveModal'; // Set Permit Holder To Inactive modal
 import SetPermitHolderToActiveModal from '@components/admin/permit-holders/modals/SetPermitHolderToActive'; // Set Permit Holder To Active modal
 import { PermitHolderToUpdateStatus } from '@tools/pages/admin/permit-holders/types'; // Type for data required in Set Permit Holder Status modal
-import GenerateReportModal from '@components/admin/permit-holders/modals/GenerateReportModal'; // Generate report modal
+import GenerateReportModal from '@components/admin/reports/GenerateReportModal'; // Generate report modal
 
 const PAGE_SIZE = 20;
 
@@ -342,7 +342,7 @@ export default function PermitHolders() {
       <GridItem colSpan={12}>
         <Flex justifyContent="space-between" alignItems="center" marginBottom="32px">
           <Text textStyle="display-xlarge">Permit Holders</Text>
-          <Button variant="outline" onClick={() => onOpenGenerateReportModal()}>
+          <Button variant="outline" onClick={onOpenGenerateReportModal}>
             Generate a Report
           </Button>
         </Flex>
@@ -491,6 +491,7 @@ export default function PermitHolders() {
       <GenerateReportModal
         isOpen={isGenerateReportModalOpen}
         onClose={onCloseGenerateReportModal}
+        page="permit-holders"
       />
     </Layout>
   );
