+++ tools/components/admin/generate-report-steps.ts
@@ -1,0 +1,4 @@
+export enum GenerateReportModalStep {
+  selectColumns = 1,
+  export = 2,
+}
